refactor: where a city is and how high it sits leaves ChunkPlan (#11)
First cut of the ChunkPlan half. CityField answers two questions at one
coordinate: whether the city mask covers it, and which of the preset's
nine level bands the terrain falls into.

One class rather than two because they are mutually recursive -
averaging the heightmap over a neighbourhood only counts the neighbours
that are themselves city - and because they are the same kind of thing:
a pure function of the world seed and a chunk coordinate, depending on no
building decision.

That last part is load-bearing and now stated where it lives. These are
consulted while the chunk candidate graph is still being computed, so
they must not read anything that depends on a building decision, its own
or a neighbour's, or the graph stops being acyclic. ChunkPlan.isCity is
the decided form and stays behind, in the layer above.

getCityLevelGui becomes cityLevelUncached, which is what it is: the same
resolution without the cache lookup. Nothing outside ChunkPlan called it
under either name.

Pure code motion, verified token-for-token: 612 tokens on both sides. All
40 external call sites were fully qualified ChunkPlan.x(, so rewiring
them carried no ambiguity.

ChunkPlan 1939 -> 1802 lines.

All seven digest suites unchanged: 688914e862e938fb (check, shuffled),
7b5348f28e0a6d23 (features), b37050817cd94b93 (avoid, expiry),
53290e1a9849c43d (avoid modes), 3fff027c14eea4a9 (rail).
